cleaned up and deleted lots of useless/old code
