+++ client/components/chat.js
@@ -1,11 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { socket } from '../socket';
-//import Button from 'react-bootstrap/lib/Button';
-//import Input from 'react-bootstrap/lib/Input';
-//import Panel from 'react-bootstrap/lib/Panel';
 import { Button, Panel } from 'react-bootstrap';
-//import { newMessage } from '../store';
+
+import { socket } from '../socket';
 
 function sendMessage() {
   let message = $('#chat-input')[0].value;
