+++ client/components/user-home.spec.js
@@ -1,13 +1,13 @@
 /* global describe beforeEach it */
 
-import {expect} from 'chai'
+import { expect } from 'chai'
 import React from 'react'
-import enzyme, {shallow} from 'enzyme'
+import enzyme, { shallow } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import {UserHome} from './user-home'
+import { UserHome } from './user-home'
 
 const adapter = new Adapter()
-enzyme.configure({adapter})
+enzyme.configure({ adapter })
 
 describe('UserHome', () => {
   let userHome
