+++ client/store/index.js
@@ -5,7 +5,6 @@ import { composeWithDevTools } from 'redux-devtools-extension'
 import me from './me';
 import messages from './messages';
 import channels from './channels';
-import currChannel from './currChannel';
 import inactiveUser from './inactiveUser';
 //import rtcConnection, {refresh} from './rtcConnection';
 
@@ -13,7 +12,6 @@ const reducer = combineReducers({
   me,
   messages,
   channels,
-  currChannel,
   inactiveUser
   //rtcConnection,
 });
@@ -29,6 +27,5 @@ export default store;
 export * from './me';
 export * from './messages';
 export * from './channels';
-export * from './currChannel';
 export * from './inactiveUser';
 //export * from './rtcConnection';
