--- utils/currentChannel.js
@@ -1,3 +1,0 @@
-export function currentChannelList(categoryName, channelList) {
-    return channelList.filter(channel => channel.category === categoryName);
-}
