--- utils/uniqueChannels.js
@@ -1,3 +1,0 @@
-export function uniqueChannelList (channelList) {
-    return Array.from(new Set(channelList.map(channel => channel.category))).sort();
-}
